src/backend/requirements.txt
Flask
Flask-Cors
biopython
celery
redis